+++ CHANGELOG.md
@@ -21,7 +21,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ---
 
 ### Fixes
-Change: "generic_db_helpers.get_current_user" use "request.user.public_id" instead of "self.request.user["public_id"]" [FA-122].
+Fix "TypeError: 'AuthTokenPayload' object is not subscriptable" error in "generic_db_helpers.get_current_user" by using "request.user.public_id" instead of "self.request.user["public_id"]" [FA-122].
 
 
 ## 0.1.5 (2024-04-20)